SWORD25: Rename lua serialization functions to use 'persist' in order to match the rest of the engine
