+++ engines/sword25/module.mk
@@ -83,9 +83,9 @@ MODULE_OBJS := \
 	util/lua/lzio.o \
 	util/lua/scummvm_file.o \
 	util/double_serializer.o \
-	util/lua_serialization_util.o \
-	util/lua_serializer.o \
-	util/lua_unserializer.o
+	util/lua_persistence_util.o \
+	util/lua_persist.o \
+	util/lua_unpersist.o
 
 # This module can be built as a plugin
 ifeq ($(ENABLE_SWORD25), DYNAMIC_PLUGIN)
