+++ engines/sword25/util/lua_persist.cpp
@@ -20,10 +20,10 @@
  *
  */
 
-#include "sword25/util/lua_serialization.h"
+#include "sword25/util/lua_persistence.h"
 
 #include "sword25/util/double_serializer.h"
-#include "sword25/util/lua_serialization_util.h"
+#include "sword25/util/lua_persistence_util.h"
 
 #include "common/stream.h"
 
@@ -55,7 +55,7 @@ static void serializeUpValue(SerializationInfo *info);
 static void serializeUserData(SerializationInfo *info);
 
 
-void serializeLua(lua_State *luaState, Common::WriteStream *writeStream) {
+void persistLua(lua_State *luaState, Common::WriteStream *writeStream) {
 	SerializationInfo info;
 	info.luaState = luaState;
 	info.writeStream = writeStream;
