+++ engines/sword25/util/lua_persistence.h
@@ -36,8 +36,8 @@ namespace Lua {
 
 #define PERMANENT_TYPE 101
 
-void serializeLua(lua_State *luaState, Common::WriteStream *writeStream);
-void unserializeLua(lua_State *luaState, Common::ReadStream *readStream);
+void persistLua(lua_State *luaState, Common::WriteStream *writeStream);
+void unpersistLua(lua_State *luaState, Common::ReadStream *readStream);
 
 } // End of namespace Lua
