+++ engines/sword25/util/lua_persistence_util.h
